+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed slow window resizing
 - Sounds playing when preset is loaded
+- Description text not updating when preset is loaded
 
 ## [1.0.0] - 2020-05-28
 ### Added